refactor: Price 값객체를 common 패키지로 분리
